backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
Pillow
easyocr
google-generativeai
python-dotenv
fpdf2
